Moved some utility functions to another folder.
