+++ part2/phonebook/db.json
@@ -19,16 +19,6 @@
       "name": "Mary Poppendieck",
       "number": "39-23-6423122",
       "id": 4
-    },
-    {
-      "id": 5,
-      "name": "Daddio",
-      "number": "502-221"
-    },
-    {
-      "id": 6,
-      "name": "John Skyler",
-      "number": "502-2210554"
     }
   ]
 }
