+++ part2/phonebook/src/App.jsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from 'react'
-import axios from 'axios';
 import phoneServices from './services/phone'
+import phoneUtils from './services/utils';
 import Filter from './components/Filter'
 import PersonForm from './components/PersonForm';
 
@@ -25,21 +25,12 @@ const App = () => {
   ? person 
   : person.name.toLowerCase().includes(filter.trim().toLowerCase()))
 
-  const nameConflict = (name) => {
-    for (let i = 0; i < persons.length; i++) {
-      if (persons[i].name === name) {
-        return true;
-      }
-    }
-    return false;
-  }
-
   const handleFormSubmit = (e) => {
     e.preventDefault();
     const name = e.target.name.value;
     const number = e.target.number.value;
-    clearFields(e);
-    if (nameConflict(name)) {
+    phoneUtils.clearFields(e);
+    if (phoneUtils.nameConflict(name, persons)) {
       alert(`${name} is already added to the phonebook`);
     } else {
       const newPerson = { id: persons.length + 1, name, number }
@@ -49,11 +40,6 @@ const App = () => {
     }
   }
 
-  const clearFields = (e) => {
-    e.target.name.value = '';
-    e.target.number.value = '';
-  }
-
   return (
     <div>
       <h2>Phonebook</h2>
@@ -61,7 +47,11 @@ const App = () => {
       <h2>add a new</h2>
       <PersonForm onFormSubmit={handleFormSubmit}/>
       <h2>Numbers</h2>
-      {peopleToShow.map(person => <p key={person.id}>{person.name} {person.number}</p>)}
+      {peopleToShow.map(person => 
+      <p key={person.id}>
+        {person.name} {person.number}
+        <button>delete</button>
+      </p>)}
     </div>
   )
 }
